+++ English Helper/English Helper/PictureGame/View/GameOptionsView.swift
@@ -30,6 +30,7 @@ struct GameOptionsView: View {
                     .pickerStyle(.segmented)
 
             }
+            Toggle("Auto Play Audio", isOn: $vm.audoPlayAudio)
             Spacer()
             Text("Select Topics")
                 .onTapGesture {
@@ -40,6 +41,8 @@ struct GameOptionsView: View {
                         ListTopicView()
                     }
                 }
+            Text("")
+            Text("")
         }.padding(.horizontal)
     }
 }
